[Entity Store] Adopt client pattern to crud
