+++ x-pack/solutions/security/plugins/entity_store/server/types.ts
@@ -24,12 +24,12 @@ import type { Logger } from '@kbn/logging';
 import type { SpacesPluginSetup, SpacesPluginStart } from '@kbn/spaces-plugin/server';
 import type { CoreSetup } from '@kbn/core/server';
 import type { ElasticsearchClient } from '@kbn/core/server';
-import type { AssetManagerClient } from './domain/asset_manager/asset_manager_client';
+import type { AssetManagerClient } from './domain/asset_manager';
 import type { EntityMaintainersClient } from './domain/entity_maintainers';
 import type { FeatureFlags } from './infra/feature_flags';
 import type { CcsLogsExtractionClient, LogsExtractionClient } from './domain/logs_extraction';
 import type { HistorySnapshotClient } from './domain/history_snapshot';
-import type { CRUDClient } from './domain/crud';
+import type { CRUDClient } from './domain/crud/crud_client';
 import type { RegisterEntityMaintainerConfig } from './tasks/entity_maintainers/types';
 
 export interface EntityStoreSetupPlugins {
